// LLMHackathon-UI_marketing/src/models/template.model.ts
export class Template {
    countryCode: string | undefined;
    countryName: string | undefined;
    stateCode: string | undefined;
    state: string | undefined;
    cityCode: number | undefined;
    cityName: string | undefined;
    code: number | undefined;
    name: string | undefined;
    rating: number | undefined;
    description: string | undefined;
    imageUrl: string | undefined;
    isActive: boolean | undefined;
}